errorMessage, check_title, check_description, check_timestamp, check_price functions moved
